+++ CHANGELOG.md
@@ -208,6 +208,8 @@
 - Docs: add NVIDIA/Cumulus/Mellanox to SNMP collector vendor list ([#22063](https://github.com/netdata/netdata/issues/22063))
 - Improve dbengine memory protection message on startup ([#22045](https://github.com/netdata/netdata/issues/22045))
 - Regenerate integrations docs ([#22070](https://github.com/netdata/netdata/issues/22070))
+- Improve ML model loading ([#22073](https://github.com/netdata/netdata/issues/22073))
+- Optimize EBPF memory handling and PID indexing. ([#22050](https://github.com/netdata/netdata/issues/22050))
 
 ## [2.9.0] - 2026-02-16